Risoluzione bug che convertiva i double in int e faceva crashare il programma. Aggiunto controllo che impedisce di mettere ";" e "," in nome e ingredienti che sennò causerebbero problemi allo split

+++ Borelli_GestionaleVacanze/Form3.cs
@@ -568,7 +568,7 @@ namespace Borelli_GestionaleVacanze
                     {
                         for (int j = i; j < listuccina.Items.Count; j++)
                         {
-                            if (int.Parse(listuccina.Items[i].SubItems[1].Text) > int.Parse(listuccina.Items[j].SubItems[1].Text))
+                            if (double.Parse(listuccina.Items[i].SubItems[1].Text) > double.Parse(listuccina.Items[j].SubItems[1].Text))
                                 ScambiaElementi(i, j, listuccina);
                         }
                     }
